requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background jobs
celery==5.3.4
redis==5.0.1

# Caching and rate limiting
aioredis==2.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# AWS SDK
boto3==1.34.0

# HTTP client
httpx==0.25.2

# Utilities
python-jose[cryptography]==3.3.0
email-validator==2.1.0
